Fix for crash on load, when running on iOS 5.
Looking at the docs for 5.0, 5.1, 6.0, and 6.1 for NSValueTransformer, it appears that the method allowsReverseTransformation: should be a class method, not an instance method. It is strange that this bug does not appear on the 5.1 - 6.1 runtimes, and only on the 5.0 runtime, but changing to the class_getClassMethod appears to resolve this issue for iOS 5, and allows running on the simulator.

This change has been tested on iOS 5.0 - 6.1 in the simulator, and on Mountain Lion (The My Mac 64 bit target).